Explicit dtypes (int) and omitting "object" dtype. Needed to load new german dataset correctly.

+++ data/processing_functions.py
@@ -167,7 +167,7 @@ def get_and_preprocess_german_short():
     categorical_col_ids = [data_columns.index(col) for col in categorical_col_names]
 
     x_values, categorical_mapping = _save_column_id_to_value_index_mapping(x_values.to_numpy(), categorical_col_ids)
-    x_values = pd.DataFrame(x_values, columns=col_names)
+    x_values = pd.DataFrame(x_values, columns=col_names, dtype=int)
     output = {
         "x_values": x_values,
         "y_values": y_values,
